requirements.txt
# Основные зависимости для интеллектуальных агентов
playwright==1.40.0
google-generativeai==0.3.2
pillow==10.1.0
python-dotenv==1.0.0

# Зависимости для AgentScope
packaging>=23.0
pydantic>=2.0.0

# Дополнительные зависимости для анализа интерфейса
inquirer>=3.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0
colorlog>=6.7.0
aiohttp>=3.8.0
typing-extensions>=4.8.0

# Зависимости для системы регистрации
faker>=30.0.0
asyncio-compat>=0.1.2
dataclasses-json>=0.6.0